log report

+++ src/shared/Shared.Entities/Entities/Twin/Report/TwinActionReport.cs
@@ -4,7 +4,7 @@ namespace Shared.Entities.Twin;
 public class TwinActionReport
 {
     public StatusType? Status { get; set; }
-    public int? Progress { get; set; }
+    public float? Progress { get; set; }
     public ResultCode? ResultCode { get; set; }
     public string ResultText { get; set; }
 }
